Feat: Add admin api and type
