+++ src/lib/Axios.ts
@@ -1,4 +1,5 @@
 import Axios from 'axios';
+import Cookies from 'js-cookie';
 
 export const tokenAxios = Axios.create({
   baseURL: process.env.NEXT_PUBLIC_DEV_API_URL,
@@ -17,3 +18,17 @@ export const axios = Axios.create({
   },
   withCredentials: true,
 });
+
+axios.interceptors.request.use(
+  (config) => {
+    const accessToken = Cookies.get('accessToken');
+    const newConfig = config;
+    if (accessToken) {
+      newConfig.headers.Authorization = `Bearer ${accessToken}`;
+    } else {
+      newConfig.headers.Authorization = 'Bearer NONE';
+    }
+    return newConfig;
+  },
+  (error) => Promise.reject(error),
+);
