+++ src/types/Admin.ts
@@ -1,0 +1,4 @@
+export type Admin = {
+  intraName: string;
+  role: string;
+};
